pyproject.toml
[tool.poetry]
name = "tech-point"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.8"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
celery = "^5.3.6"
redis = "^5.0.1"
qrcode = {extras = ["pil"], version = "^7.4.2"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["src", "tech_point"]
line_length = 79

[tool.black]
line_length = 79
